Integrate webservice with CLI + minor updates
